Move httputils to pkg dir2
