+++ pkg/httputils/DownloadAsFileOptions.go
