+++ pkg/httputils/GenericResponse.go
